pyproject.toml
[project]
name = "debating-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=6.8.0",
    "groq>=1.0.0",
    "ipython>=9.10.0",
    "python-dotenv>=1.2.1",
    "rich>=14.3.3",
]
